workspace/assignments/03-LangChain-Graph Agent/pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = "LangGraph ReAct Trading Agent with MCP Tools"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp[cli]>=1.0.0",
    "streamlit>=1.40.0",
    "qdrant-client>=1.12.0",
    "yfinance>=0.2.40",
    "tavily-python>=0.5.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "plotly>=5.24.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
